refactor createMovie function
