+++ hvpy/__init__.py
@@ -2,5 +2,6 @@ from .config import set_api_url
 from .datasource import *
 from .event import *
 from .facade import *
-from .utils import create_layers, create_events
+from .helper import createMovie
+from .utils import create_events, create_layers, save_file
 from .version import __version__
